refactor(wiki-compare): split 3.6k-line monolith into modules
scripts/wiki-compare.ts is now a thin entry point (re-exports the existing
public/test API and runs the CLI when executed directly); the implementation
moves to scripts/wiki-compare/:

- types.ts     shared types, priorities, constants
- cache.ts     API/wiki caches + results paths (gitignored data/)
- overlay.ts   override/suppression filtering
- normalize.ts text/item/map normalization and fuzzy matching
- api.ts       tarkov.dev fetching + task/wiki-title resolution
- wiki.ts      MediaWiki fetch + wikitext parsing
- compare.ts   API-vs-wiki comparison
- cli.ts       arg parsing and single/bulk runners

No behavior change: typecheck + full suite green, CLI verified end-to-end
(--help and a live single-task comparison).
